# docs/reference/openapi.yaml
...
openapi: 3.1.0
info:
  title: Narad HTTP API
  version: v1
  summary: Produce, consume and manage topics on a Narad cluster over HTTP.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  x-narad-lead: |
    Look up every Narad HTTP endpoint: its parameters, request body, the
    grant it needs, the status codes it returns and a real example.
  x-narad-lead-example:
    request: |
      curl -i -u "$AUTH" "$NARAD/v1/topics/orders?partition=1"
    response: |
      HTTP/1.1 200 OK
      Content-Length: 456
      Content-Type: application/json
      Date: Mon, 28 Sep 2026 19:31:43 GMT

      {
        "name": "orders",
        "id": "681a429c0b683b2d",
        "partitions": 3,
        "retention_ms": 172800000,
        "visibility_timeout_ms": 30000,
        "max_in_flight_per_partition": 1024,
        "max_acked_ahead_per_partition": 1024,
        "created_at": 1790623903,
        "owner": "admin",
        "role": "parent",
        "children": ["orders-audit"],
        "schema_version": 0,
        "partition_stats": [
          {
            "index": 1,
            "segments": 1,
            "oldest_offset": 0,
            "next_offset": 2,
            "high_watermark": 2,
            "size_bytes": 166,
            "oldest_segment_at": 1790623903,
            "owner_node": "narad-0"
          }
        ]
      }
  x-narad-placeholders: |
    - `$NARAD` is the base URL of any node or of the load balancer in front
...
      off and `-u "$AUTH"` can be left out.

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - basicAuth: []
...
paths:
...
  /v1/topics/{topic}:
    parameters:
      - $ref: '#/components/parameters/TopicPath'
    get:
      operationId: getTopic
      tags: [Topics]
      summary: Get a topic
      description: |
        Returns the topic, its current schema and statistics for each
        partition. The node that answers asks every partition's owner for
        its statistics.

        **New in v3.1.0:** when some partitions' owners are down, the answer
        is still `200`. The partitions that could be read carry their
        statistics and `status` `ok`; each other one is a placeholder with
        zero statistics, `status` `owner_unavailable` and the owner's
        `owner_liveness`, and the body carries `partial: true`. Leave the
        placeholders out of any total. An owner that does not answer within
        2 seconds counts as `unreachable`. A v3.0.1 node answers `421`
        instead.

        **Unreleased:** a [remote child](glossary.md#remote-child)'s stub
        has `partitions: 0` and a `remote` object that names the remote and
        the topic there. Its `created_by` and `paused_by` are shown to
        admins only.
      x-narad-grant: |
        Any grant that matches the topic name, ownership, or `admin`.
      parameters:
        - name: partition
          in: query
          description: Return statistics for this partition only.
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: The topic with partition statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicDetails'
        '400':
          description: '`partition` is not a partition of the topic.'
        '401':
          description: Missing or wrong credentials.
        '403':
          description: No grant on the topic.
        '404':
          description: The topic does not exist.
        '421':
          description: From a v3.0.1 node, the owner of one of the topic's partitions could not be found or refused to answer. An upgraded node answers `200` with `partial` instead.
        '503':
          description: The answering node could not read its own copy of the cluster metadata (for example while it catches up after a restart). Retry.
      x-narad-examples:
        - request: |
            curl -i -u "$AUTH" "$NARAD/v1/topics/orders?partition=1"
          response: |
            HTTP/1.1 200 OK
            Content-Length: 470
            Content-Type: application/json
            Date: Mon, 28 Sep 2026 19:31:43 GMT

            {"name":"orders","id":"681a429c0b683b2d","partitions":3,"retention_ms":172800000,"visibility_timeout_ms":30000,"max_in_flight_per_partition":1024,"max_acked_ahead_per_partition":1024,"created_at":1790623903,"owner":"admin","role":"parent","children":["orders-audit"],"schema_version":0,"partition_stats":[{"index":1,"segments":1,"oldest_offset":0,"next_offset":2,"high_watermark":2,"size_bytes":166,"oldest_segment_at":1790623903,"owner_node":"narad-0","status":"ok"}]}
    patch:
      operationId: alterTopic
      tags: [Topics]
      summary: Change a topic
      description: |
        Changes retention, the per-partition caps or the partition count, or
        registers a new schema version. Send at least one field.
        `visibility_timeout_ms` is fixed when the topic is created.

        Fields are applied one group at a time in this order: retention,
        then the caps, then partitions, then schema. There is no
        transaction across them: the first failure answers its error and
        the groups before it stay applied. Send one field per request when
        you need all or nothing.
      x-narad-grant: |
        Ownership of the topic, or `admin`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlterTopicRequest'
      responses:
        '200':
          description: Changed. The body is the topic after the change.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
        '400':
          description: No field given, a value is invalid, `partitions` is not larger than the current count, or the new schema is not compatible.
        '401':
          description: Missing or wrong credentials.
        '403':
          description: Not the owner and not `admin`, as the node that answers or the cluster leader sees the topic.
        '404':
          description: The topic does not exist, also after the answering node caught up with the leader.
        '409':
          description: '`schema_base_version` is not the current version, the history holds 1000 versions or (from v3.1.0) the new version would take it past 4 MiB or the cluster''s schemas past 256 MiB, the topic is a child whose schema its parent manages, the new retention is too short for a delay child, or (from v3.1.0) the topic was deleted and recreated, or grew, under the request twice in a row (`topic changed since it was read`). Unreleased: the topic is a remote child''s stub, which takes no change, or the new retention is below 24 hours on a parent with remote children (a retention already below it may still grow).'
        '413':
          description: The body is over 1 MiB.
        '415':
          description: No accepted `Content-Type` and no `X-Narad-Client` header.
        '503':
          description: The cluster has no leader to write the change, the answering node could not reach the leader to confirm a topic it does not have, or (from v3.1.0) a partition increase found every live node being decommissioned, so no node can take the new partitions; nothing was changed.
      x-narad-examples:
        - request: |
            curl -i -u "$AUTH" -X PATCH "$NARAD/v1/topics/orders" \
              -H "Content-Type: application/json" \
              -d '{"retention_ms": 172800000}'
          response: |
            HTTP/1.1 200 OK
            Content-Length: 244
            Content-Type: application/json
            Date: Mon, 28 Sep 2026 19:31:43 GMT

            {"name":"orders","id":"681a429c0b683b2d","partitions":3,"retention_ms":172800000,"visibility_timeout_ms":30000,"max_in_flight_per_partition":1024,"max_acked_ahead_per_partition":1024,"created_at":1790623903,"owner":"admin","role":"standalone"}
    delete:
      operationId: deleteTopic
      tags: [Topics]
      summary: Delete a topic
      description: |
        Deletes the topic's metadata and its data on every node, and
        removes its fan-out links. There is no undo. Once the metadata
        delete is committed the answer is `204`, even if a node that is
        down could not remove its files yet; that node removes them when
        it next starts.

        **Unreleased:** deleting a parent that has
        [remote children](glossary.md#remote-child) deletes their stubs in
        the same entry, and deleting a remote child's stub deletes the link.
        Either is refused with `409` while any record of the parent is not
        yet on the remote, unless `force=true` abandons them; the leader
        checks every member's ingress backlog and every cursor first. The
        same refusal, with its body, is in
        [Detach a child](#detach-child).
      x-narad-grant: |
        Ownership of the topic, or `admin`. Unreleased: a remote child's
        stub has no owner; the owner of its parent, or an `admin`, deletes
        it, with security on.
      parameters:
        - name: force
          in: query
          x-narad-since: unreleased
          description: |
            `true` deletes a parent with remote children, or a remote
            child's stub, even while records are unshipped, abandoning
            them. Ignored for any other topic. `narad topic rm --force`
            never sends it: there `--force` only skips the prompt.
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Deleted.
        '400':
          description: Unreleased. `force` is not `true` or `false`.
        '401':
          description: Missing or wrong credentials.
        '403':
          description: Not the owner and not `admin`, as the node that answers or the cluster leader sees the topic. For a remote child's stub (unreleased), the caller is neither an `admin` nor the parent's owner, or security is off. For a parent with remote children (unreleased), security is off (`remotes require security`).
        '404':
          description: The topic does not exist, also after the answering node caught up with the leader.
        '409':
          description: New in v3.1.0. The topic was deleted and recreated under the request twice in a row (`topic changed since it was read`); nothing was deleted. Read the topic again before deleting it. Unreleased, for a parent with remote children or a stub without `force`, records of the parent are not yet on the remote (the body carries `lag_messages`, `lag_complete` and `dispatch_backlog`), or the leader found a remote child the answering node did not know of yet; retry.
        '412':
          description: Unreleased. The cluster leader runs a release without remote children; finish the upgrade.
        '429':
          description: Unreleased. The leader ran an unshipped check for this parent less than 10 seconds ago; retry after `Retry-After`.
          x-narad-origin: forwarding
        '501':
          description: Unreleased. The answering node has no remote plane, so it cannot delete a remote-linked topic.
        '503':
          description: The cluster has no leader to write the delete, or the answering node could not reach the leader to confirm a topic it does not have. Unreleased, for a remote-linked topic, also when the unshipped check could not run.
      x-narad-examples:
        - request: |
            curl -i -u "$AUTH" -X DELETE "$NARAD/v1/topics/orders-audit"
          response: |
            HTTP/1.1 204 No Content
            Date: Mon, 28 Sep 2026 19:31:43 GMT
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
    TopicPath:
      name: topic
      in: path
      required: true
      description: Name of the topic.
      schema:
        type: string
...
  schemas:
...
    AlterTopicRequest:
      type: object
      additionalProperties: false
      properties:
        retention_ms:
          type: integer
          description: New retention. At least 3,600,000 (1 hour), or `0` to keep messages forever; a negative value gets `400`. To return to the server default, send its value.
        max_in_flight_per_partition:
          type: integer
          description: New in-flight cap; `0` sets the server default. A cap left out keeps its value.
        max_acked_ahead_per_partition:
          type: integer
          description: New acked-ahead cap; `0` sets the server default. A cap left out keeps its value.
        partitions:
          type: integer
          description: New partition count, larger than the current one and at most the server's maximum (108 by default). New keys may then map to other partitions.
        schema:
          description: |
            A new schema version, checked for compatibility with the current
            one. Sending the current schema again changes nothing and
            answers `200`. A schema cannot be removed (`null` gets `400`).
        schema_base_version:
          type: integer
          description: With `schema`, apply it only if the current version is exactly this number; `409` otherwise.
    Topic:
      type: object
      properties:
        name:
          type: string
          description: Topic name.
        id:
          type: string
          description: The [incarnation](glossary.md#incarnation) ID, 16 hex characters. A topic deleted and created again under the same name gets a new one. Absent on a topic created before v2.2.0.
        partitions:
          type: integer
          description: Partition count.
        retention_ms:
          type: integer
          description: Retention in milliseconds; `0` keeps messages forever.
        visibility_timeout_ms:
          type: integer
          description: Lease length in milliseconds.
        max_in_flight_per_partition:
          type: integer
          description: In-flight cap per partition.
        max_acked_ahead_per_partition:
          type: integer
          description: Acked-ahead cap per partition.
        created_at:
          type: integer
          description: Creation time, Unix seconds.
        owner:
          type: string
          description: The user that created the topic. Absent when security was off.
        role:
          type: string
          enum: [standalone, parent, child]
          description: Fan-out role. Absent in the answer to a create without `parent`, which means `standalone`.
        children:
          type: array
          items:
            type: string
          description: A parent's children, in attach order.
        parent:
          type: string
          description: A child's parent.
        attach_epoch:
          type: string
          description: A child's current attachment; it changes on every attach.
        fanout_delay_ms:
          type: integer
          description: A delay child's delay in milliseconds.
        attach_offsets:
          type: array
          items:
            type: integer
          description: A child's attach point, one offset per parent partition.
        remote:
          $ref: '#/components/schemas/RemoteLink'
    RemoteLink:
      type: object
      x-narad-since: unreleased
      description: A [remote child](glossary.md#remote-child)'s link. Present only on a remote child's stub, which has `partitions` `0` and no owner.
      properties:
        name:
          type: string
          description: The remote the copies go to.
        topic:
          type: string
          description: The topic on the remote.
        target_id:
          type: string
          description: The target topic's ID as the attach, or the last resume with `accept_target`, saw it. A target on v3.1.0 serves it in its describe answer, so recreate detection works there too. Empty for a target topic created before topic IDs (v2.1 and earlier); such a link stops in `target_replaced` if the target later reports an ID, because a topic gains one only by being recreated.
        from:
          type: string
          enum: [attach, unconsumed, earliest]
          description: Where the link started on each parent partition.
        lanes:
          type: integer
          description: Ordered streams per parent partition, 1 to 8.
        paused:
          type: boolean
          description: '`true` while paused. Absent otherwise.'
        pause_reason:
          type: string
          description: The reason given to pause.
        paused_by:
          type: string
          description: The admin who paused it; shown to admins only.
        paused_at_ms:
          type: integer
          description: When it was paused, Unix milliseconds.
        skip:
          type: object
          additionalProperties:
            type: array
            items:
              type: integer
          description: Per parent partition, the offsets an admin accepted to lose, ascending, at most 4000. A cursor drops a record only while it is stuck on exactly one of them.
        created_by:
          type: string
          description: The admin who attached it; shown to admins only.
    TopicDetails:
      allOf:
        - $ref: '#/components/schemas/Topic'
        - type: object
          properties:
            schema_version:
              type: integer
              description: Current schema version, `0` without a schema.
            schema:
              description: The current schema document. Absent without a schema.
            partition_stats:
              type: array
              description: One entry per partition, or only the one `partition` asked for.
              items:
                $ref: '#/components/schemas/PartitionStats'
            partial:
              type: boolean
              x-narad-since: v3.1.0
              description: '`true` when some entries are placeholders whose owner could not report (`status` `owner_unavailable`). Absent otherwise.'
    PartitionStats:
      type: object
      properties:
        index:
          type: integer
          description: Partition number.
        segments:
          type: integer
          description: Segment files on disk.
        oldest_offset:
          type: integer
          description: Lowest offset still kept.
        next_offset:
          type: integer
          description: Offset the next record will get. It can lead `high_watermark` while a commit runs.
        high_watermark:
          type: integer
          description: One past the last offset consumers can see.
        size_bytes:
          type: integer
          description: Bytes on disk.
        oldest_segment_at:
          type: integer
          description: Time of the oldest segment, Unix seconds. Absent when unknown.
        owner_node:
          type: string
          description: ID of the node that owns the partition.
        status:
          type: string
          enum: [ok, owner_unavailable]
          x-narad-since: v3.1.0
          description: '`ok` when the statistics are the owner''s; `owner_unavailable` for a placeholder with zero statistics, because the owner could not report them. Never add a placeholder''s numbers to a total.'
        owner_liveness:
          type: string
          enum: [dead, unreachable, unknown, unassigned]
          x-narad-since: v3.1.0
          description: 'Why an `owner_unavailable` partition''s owner could not report: `dead` (marked dead), `unreachable` (alive, but its statistics did not come back within 2 seconds), `unknown` (no member with an address), `unassigned` (no owner yet). Absent for `ok`.'
...